rename: Tool -> Shape 이름 변경
