+++ app/src/main/java/woowacourse/paint/Painting.kt
@@ -4,22 +4,22 @@ import android.graphics.Canvas
 import android.graphics.Paint
 import android.graphics.PorterDuff
 import android.graphics.PorterDuffXfermode
-import woowacourse.paint.tool.Tool
+import woowacourse.paint.shape.Shape
 
 class Painting(
-    val tool: Tool,
+    val shape: Shape,
     val paint: Paint,
 ) {
     fun startDraw(pointX: Float, pointY: Float) {
-        tool.startDraw(pointX, pointY)
+        shape.startDraw(pointX, pointY)
     }
 
     fun onDraw(pointX: Float, pointY: Float) {
-        tool.onDraw(pointX, pointY)
+        shape.onDraw(pointX, pointY)
     }
 
     fun drawPath(canvas: Canvas) {
-        tool.drawPath(canvas, paint)
+        shape.drawPath(canvas, paint)
     }
 
     fun changeSize(value: Float) {
@@ -31,15 +31,15 @@ class Painting(
     }
 
     fun copy(): Painting {
-        return Painting(tool.copy(), Paint(paint))
+        return Painting(shape.copy(), Paint(paint))
     }
 
-    fun changeTool(tool: Tool, style: Paint.Style, isEraser: Boolean = false): Painting {
+    fun changeShape(shape: Shape, style: Paint.Style, isEraser: Boolean = false): Painting {
         when (isEraser) {
             true -> paint.xfermode = PorterDuffXfermode(PorterDuff.Mode.CLEAR)
             else -> paint.xfermode = null
         }
         paint.style = style
-        return Painting(tool, paint)
+        return Painting(shape, paint)
     }
 }
