+++ app/src/main/java/woowacourse/paint/Tools.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint.tool
+package woowacourse.paint
 
 enum class Tools {
     LINE, RECTANGLE, CIRCLE, ERASER
