+++ app/src/main/java/woowacourse/paint/shape/Shape.kt
@@ -1,10 +1,10 @@
-package woowacourse.paint.tool
+package woowacourse.paint.shape
 
 import android.graphics.Canvas
 import android.graphics.Paint
 
-interface Tool {
-    fun copy(): Tool
+interface Shape {
+    fun copy(): Shape
     fun startDraw(pointX: Float, pointY: Float)
     fun onDraw(pointX: Float, pointY: Float)
     fun drawPath(canvas: Canvas, paint: Paint)
